refactor: unify copy to clip component
