+++ src/modules/common/components/CopyToClip/index.ts
@@ -1,0 +1,1 @@
+export * from './CopyToClip';
